Refactor models file structure
