+++ backend/src/main/java/gr/uom/strategicplanning/models/domain/Commit.java
@@ -1,4 +1,4 @@
-package gr.uom.strategicplanning.models;
+package gr.uom.strategicplanning.models.domain;
 
 import lombok.*;
