+++ backend/src/main/java/gr/uom/strategicplanning/models/domain/Developer.java
@@ -1,4 +1,4 @@
-package gr.uom.strategicplanning.models;
+package gr.uom.strategicplanning.models.domain;
 
 import lombok.*;
 
@@ -13,24 +13,15 @@ import javax.persistence.ManyToOne;
 @AllArgsConstructor
 @NoArgsConstructor
 @ToString
-public class Language {
-    
+public class Developer {
+
     @Id
     @GeneratedValue
     private Long id;
     private String name;
-    private String version;
-    private String imageUrl;
-    private Number linesOfCode;
-
+    private String githubUrl;
+    private int totalCommits;
     @ManyToOne
     private Project project;
 
-    public Language(String name) {
-        this.name = name;
-    }
-
-    public boolean is(String language) {
-        return this.name.equals(language);
-    }
 }
