+++ backend/src/main/java/gr/uom/strategicplanning/models/domain/Language.java
@@ -1,4 +1,4 @@
-package gr.uom.strategicplanning.models;
+package gr.uom.strategicplanning.models.domain;
 
 import lombok.*;
 
@@ -13,15 +13,24 @@ import javax.persistence.ManyToOne;
 @AllArgsConstructor
 @NoArgsConstructor
 @ToString
-public class Developer {
-
+public class Language {
+    
     @Id
     @GeneratedValue
     private Long id;
     private String name;
-    private String githubUrl;
-    private int totalCommits;
+    private String version;
+    private String imageUrl;
+    private Number linesOfCode;
+
     @ManyToOne
     private Project project;
 
+    public Language(String name) {
+        this.name = name;
+    }
+
+    public boolean is(String language) {
+        return this.name.equals(language);
+    }
 }
