+++ backend/src/main/java/gr/uom/strategicplanning/models/stats/ActivityStats.java
@@ -1,4 +1,4 @@
-package gr.uom.strategicplanning.models;
+package gr.uom.strategicplanning.models.stats;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -14,15 +14,15 @@ import javax.persistence.Id;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class CodeSmell {
-
+public class ActivityStats {
+    
     @Id
     @GeneratedValue
     private Long id;
-    private String name;
-    private String severityLevel;
-    private int line;
-    private int remediationTime;
-    private String description;
-
+    private float commitsPerDay;
+    private float locAddedPerDay;
+    private float filesAddedPerDay;
+    private float projectsAddedPerDay;
+    private float averageLoC;
+    
 }
